Add testing setup
