+++ package.json
@@ -7,7 +7,8 @@
     "build": "ncc build src/main.ts",
     "format": "prettier --write **/*.ts",
     "format-check": "prettier --check **/*.ts",
-    "test": "jest"
+    "test": "jest",
+    "test:watch": "jest --watchAll"
   },
   "repository": {
     "type": "git",
@@ -27,9 +28,13 @@
     "@actions/github": "^4.0.0"
   },
   "devDependencies": {
+    "@types/jest": "^26.0.20",
     "@types/node": "^14.14.34",
     "@vercel/ncc": "^0.27.0",
+    "jest": "^26.6.3",
+    "nock": "^13.0.11",
     "prettier": "^2.2.1",
+    "ts-jest": "^26.5.3",
     "typescript": "^4.2.3"
   }
 }
