+++ src/main.ts
@@ -1,28 +1,10 @@
 import * as core from "@actions/core";
 import * as github from "@actions/github";
+import { approve } from "./approve";
 
 async function run() {
-  try {
-    const token = core.getInput("github-token", { required: true });
-
-    const { pull_request: pr } = github.context.payload;
-    if (!pr) {
-      throw new Error("Event payload missing `pull_request`");
-    }
-
-    const client = github.getOctokit(token);
-
-    core.debug(`Creating approving review for pull request #${pr.number}`);
-    await client.pulls.createReview({
-      owner: github.context.repo.owner,
-      repo: github.context.repo.repo,
-      pull_number: pr.number,
-      event: "APPROVE",
-    });
-    core.debug(`Approved pull request #${pr.number}`);
-  } catch (error) {
-    core.setFailed(error.message);
-  }
+  const token = core.getInput("github-token", { required: true });
+  await approve(token, github.context.payload);
 }
 
 run();
