+++ tsconfig.json
@@ -8,5 +8,5 @@
     "noImplicitAny": false,
     "esModuleInterop": true
   },
-  "exclude": ["node_modules", "**/*.test.ts"]
+  "exclude": ["node_modules"]
 }
